modify: address-apis

+++ src/app/Address/addressController.js
@@ -13,7 +13,7 @@ const kakaoMap = require("../../../controllers/kakao_ctrl").getAddressInfo;
 
 /**
  * API No. 5
- * API Name : 주소 추가 API
+ * API Name : 주소 추가 API (추가와 동시에 사용자 위치 변경)
  * [POST] /addresses
  */
 exports.createAddresses = async function (req, res) {
@@ -89,11 +89,19 @@ exports.modifyAddresses = async function (req, res) {
   if (checkAddressExist === 0)
     return res.send(errResponse(baseResponse.ADDRESS_IS_NOT_EXIST)); // 3007
 
+  const checkAddressDeleted = await addressProvider.checkAddressDeleted(
+    addressId
+  );
+
+  if (checkAddressDeleted === 0)
+    return res.send(errResponse(baseResponse.ADDRESS_IS_DELETED)); // 3009
+
   // Response Error End
 
   const { lat, lng } = await kakaoMap(address);
 
   const result = await addressService.updateAddress(
+    userId,
     type,
     nickname,
     buildingName,
@@ -136,6 +144,13 @@ exports.deleteAddresses = async function (req, res) {
   if (checkAddressExist === 0)
     return res.send(errResponse(baseResponse.ADDRESS_IS_NOT_EXIST)); // 3007
 
+  const checkAddressDeleted = await addressProvider.checkAddressDeleted(
+    addressId
+  );
+
+  if (checkAddressDeleted === 0)
+    return res.send(errResponse(baseResponse.ADDRESS_IS_DELETED)); // 3009
+
   // Response Error End
 
   const result = await addressService.deleteAddress(addressId);
@@ -170,3 +185,63 @@ exports.selectAddresses = async function (req, res) {
 
   return res.send(response(baseResponse.SUCCESS, result));
 };
+
+/**
+ * API No. 16
+ * API Name : 집/회사 주소 존재 여부 확인 API
+ * [GET] /addresses/:type/house-company
+ */
+exports.checkHouseCompany = async function (req, res) {
+  const { userId } = req.verifiedToken;
+
+  const { type } = req.params;
+
+  // Request Error Start
+
+  if (!userId) return res.send(errResponse(baseResponse.USER_ID_IS_EMPTY)); // 2010
+
+  // Request Error End
+
+  // Response Error Start
+
+  const checkUserExist = addressProvider.checkUserExist(userId);
+
+  if (checkUserExist === 0)
+    return res.send(errResponse(baseResponse.USER_IS_NOT_EXIST)); // 3006
+
+  // Response Error End
+
+  const result = await addressProvider.checkHouseCompany(userId, type);
+
+  return res.send(response(baseResponse.SUCCESS, { type, exist: result }));
+};
+
+/**
+ * API No. 17
+ * API Name : 주소 목록에서 주소 선택 API
+ * [PATCH] /addresses/user-address
+ */
+// exports.changeLocation = async function (req, res) {
+//   const { userId } = req.verifiedToken;
+
+//   const body
+
+//   // Request Error Start
+
+//   if (!userId) return res.send(errResponse(baseResponse.USER_ID_IS_EMPTY)); // 2010
+
+//   // Request Error End
+
+//   // Response Error Start
+
+//   const checkUserExist = addressProvider.checkUserExist(userId);
+
+//   if (checkUserExist === 0)
+//     return res.send(errResponse(baseResponse.USER_IS_NOT_EXIST)); // 3006
+
+//   // Response Error End
+
+//   const result = await addressService.updateLocation(userId);
+
+//   return res.send(response(baseResponse.SUCCESS, result));
+// };
